\name{AdhesiveFormulationK}
\alias{AdhesiveFormulationK}
\docType{data}
\title{
Adhesive Formulation K Data
}
\description{
A dataset from Xie al. (2015) Strength was tested at 40, 50, and 60 degree C.
}
\usage{data(AdhesiveFormulationK)}
\format{
  A data frame with 101 observations on the following 3 variables.
  \describe{
    \item{\code{TempC}}{Temperature.}
    \item{\code{TimeH}}{Time in hours.}
    \item{\code{Response}}{Strength (Newtons).}
  }
}
\source{Y. Xie, C. B. King, Y. Hong, and Q. Yang (2015). Semiparametric Models for Accelerated Destructive Degradation Test Data Analysis. Preprint: arXiv:1512.03036.

}
