\name{PolymerY}
\alias{PolymerY}
\docType{data}
\title{
Polymeric Material
}
\description{
A dataset from Tsai et al. (2013) conducting a study on the tensile strength of a new type of polymer material.
}
\usage{data(PolymerY)}
\format{
  A data frame with 76 observations on the following 3 variables.
  \describe{
    \item{\code{TempC}}{Temperature.}
    \item{\code{TimeH}}{Time in hours.}
    \item{\code{Response}}{Strength (Newtons).}
  }
}
\source{Tsai, C.-C., S.-T. Tseng, N. Balakrishnan, and C.-T. Lin (2013). Optimal design for accelerated destructive degradation tests. \emph{Quality Technology and Quantitative Management 10, 263-276.}
}
