\name{SealStrength}
\alias{SealStrength}
\docType{data}
\title{
Data for Seal Strength
}
\description{
Seal Strength dataset presented in Li and Doganaksoy (2014).
}
\usage{data(SealStrength)}
\format{
  A data frame with 210 observations on the following 3 variables.
  \describe{
    \item{\code{TempC}}{Temperature.}
    \item{\code{TimeH}}{Time in hours.}
    \item{\code{Response}}{Strength (Newtons)}
  }
}
\details{
Seal strength dataset is collected under an experiment of testing the strength of a new seal.
}
\source{
Li, M. and N. Doganaksoy (2014). Batch variability in accelerated-degradation testing. \emph{Journal of Quality Technology 46(2)}, 171--180.
}

