\name{AdhesiveBondB}
\alias{AdhesiveBondB}
\docType{data}
\title{
Adhesive Bond B dataset
}
\description{
A dataset from Escobar et al. (2003) containing the results of an accelerated destructive degradation testing on the strength of an adhesive bond.
}
\usage{data(AdhesiveBondB)}
\format{
  A data frame with 82 observations on the following 3 variables.
  \describe{
    \item{\code{TempC}}{Temperature.}
    \item{\code{TimeH}}{Time in hours.}
    \item{\code{Response}}{Strength (Newtons).}
  }
}
\source{
L. A. Escobar, W. Q. Meeker, D. L. Kugler, and L. L. Kramer, ``Accelerated destructive degradation
tests: Data, models, and analysis,'' in Mathematical and Statistical Methods in Reliability,
B. H. Lindqvist and K. A. Doksum, Eds. River Edge, NJ: World Scientific Publishing
Company, 2003, ch. 21.
}
